chore: adjust spore_extension_lua and tests
