+++ deployment/migration/cluster/2023-12-30-041546.json
@@ -2,10 +2,10 @@
   "cell_recipes": [
     {
       "name": "cluster",
-      "tx_hash": "0xf0cc101c2b71d7e9fba2615c59d9a65036e02dcd3647c957f4fcdcb8e25155c8",
+      "tx_hash": "0x91555ebd2bbddf80da141a58220648d6dd3f8c04ba4bb8312d01907165b636ad",
       "index": 0,
       "occupied_capacity": 3764600000000,
-      "data_hash": "0xaab72c639657930e89890c25eec4ee3fcb40bb3d3488d73f80fa16a694eec911",
+      "data_hash": "0x1dbd4b03f1cf9d27b5359f8e3569efcb848bc71e6f204654509ce67c07bb1fdd",
       "type_id": "0xd70ad85c0b3c377b3f9d73f53e110de6c853139340e12aa667bba9db749ae876"
     }
   ],
