+++ deployment/migration/cluster_agent/2023-12-30-041657.json
@@ -2,10 +2,10 @@
   "cell_recipes": [
     {
       "name": "cluster_agent",
-      "tx_hash": "0xdd5145d3e8e9f2f112429e6a619acc8cd347a99d72de84d1257d9b6dc06ea304",
+      "tx_hash": "0x6d41f6921410c5227c50c8d6b99583a587f6363e917ecd1805d0c592c6560f12",
       "index": 0,
       "occupied_capacity": 4961400000000,
-      "data_hash": "0xf088c9c1513fb904e29099c29c48955ed876b83be3366caf4d5e89d65c080825",
+      "data_hash": "0x98bfdda3000708d2b058e0bd17827b6d3ca21827f059baeea23b228b74c678b3",
       "type_id": "0xbfe31502db747a6a90e20d88c7a3819e339f78200691b11057323fbdf85754fe"
     }
   ],
